MCP/Indexer/requirements.txt
fastmcp>=0.3.0
python-dotenv>=1.0.0
neo4j>=5.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
